requirements.txt
numpy
Pillow
boto3
gpiozero
picamera
pandas
opencv-python
imutils